Remove extraneous pages tests(#128)
